Bugfix: prevent badge from being positioned at wrong place
When the search bar was initialized with no filters (nrSelectedFilters = 0): the badge would be hidden and thereby not positioned next to the filter button. Onc
e a filter was selected (nrSelectedFilters > 0), the badge became visible but wa
sn't positioned correctly.
This is avoided by positioning the badge from the beginning but just making it invisible.